[Security Solution][Investigations] - Fix browser fields cache (#234381)
## Summary

When testing the managed dataViews pr here:
github.com/elastic/kibana/pull/223451 @christineweng made a great catch
that while caching dataViews based on the index pattern may seem
beneficial to avoid caching the same set of fields over and over again,
two dataViews can share the same index pattern list, but have a
different field list based on any runtime fields set on the dataView
instance. Given this, the safest approach is to just rely on the
`DataViewId` to cache the browser fields.

Expectations / Concessions:

1. AdHoc dataViews will not be cached
2. The same index pattern list for two different dataViews will now
cache two different browserField objects
3. To not overload the browser memory, we will set a cacge size limit of
10.

(cherry picked from commit 9d9e643ad08b86549883c538fe6f13b2562019e9)

# Conflicts:
#	x-pack/solutions/security/plugins/security_solution/public/cases/components/case_events/table.tsx
#	x-pack/solutions/security/plugins/security_solution/public/data_view_manager/components/data_view_picker/index.tsx
#	x-pack/solutions/security/plugins/security_solution/public/data_view_manager/hooks/use_browser_fields.ts
#	x-pack/solutions/security/plugins/security_solution/public/data_view_manager/utils/security_browser_fields_manager.ts
#	x-pack/solutions/security/plugins/security_solution/public/timelines/components/fields_browser/index.tsx
