+++ x-pack/solutions/security/plugins/security_solution/public/common/containers/source/index.tsx
@@ -12,8 +12,7 @@ import type { DataViewBase } from '@kbn/es-query';
 import type { BrowserFields } from '@kbn/timelines-plugin/common';
 import type { FieldSpec, IIndexPatternFieldList } from '@kbn/data-views-plugin/common';
 import type { DataViewSpec } from '@kbn/data-views-plugin/public';
-
-import { browserFieldsManager } from '../../../data_view_manager/utils/security_browser_fields_manager';
+import { buildBrowserFields } from '../../../data_view_manager/utils/build_browser_fields';
 import { useKibana } from '../../lib/kibana';
 import * as i18n from './translations';
 import { useAppToasts } from '../../hooks/use_app_toasts';
@@ -115,7 +114,7 @@ export const useFetchIndex = (
           abortCtrl.current = new AbortController();
           const dv = await data.dataViews.create({ title: iNames.join(','), allowNoIndex: true });
           const dataView = dv.toSpec();
-          const { browserFields } = browserFieldsManager.getBrowserFields(dv);
+          const browserFields = buildBrowserFields(dv?.fields);
 
           previousIndexesName.current = dv.getIndexPattern().split(',');
