+++ x-pack/solutions/security/plugins/security_solution/public/data_view_manager/hooks/use_browser_fields.ts
@@ -10,7 +10,7 @@ import type { BrowserFields } from '@kbn/timelines-plugin/common';
 import type { DataView } from '@kbn/data-views-plugin/common';
 import { DataViewManagerScopeName } from '../constants';
 import { useDataView } from './use_data_view';
-import { browserFieldsManager } from '../utils/security_browser_fields_manager';
+import { buildBrowserFields } from '../utils/build_browser_fields';
 import { useIsExperimentalFeatureEnabled } from '../../common/hooks/use_experimental_features';
 
 const emptyFields = {};
@@ -31,7 +31,7 @@ export const useBrowserFields = (
       return emptyFields;
     }
 
-    const { browserFields } = browserFieldsManager.getBrowserFields(activeDataView, scope);
+    const browserFields = buildBrowserFields(activeDataView?.fields);
 
     return browserFields;
   }, [activeDataView, scope]);
